# docs-site/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Overlay API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/audit": {
      "get": {
        "operationId": "getAdminAudit",
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminAuditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuditResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/app/subscription": {
      "get": {
        "operationId": "getAppSubscription",
        "tags": [
          "app"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSubscriptionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/checkout": {
      "post": {
        "operationId": "postCheckout",
        "tags": [
          "checkout"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/checkout/verify": {
      "post": {
        "operationId": "postCheckoutVerify",
        "tags": [
          "checkout"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutVerifyResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/convex/{type}": {
      "post": {
        "operationId": "postConvexType}",
        "tags": [
          "convex"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvexProxyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvexProxyResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AdminAuditRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "AdminAuditResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
      "AppSubscriptionRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "AppSubscriptionResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
      "CheckoutRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "CheckoutResponse": {
        "type": "object",
        "additionalProperties": true
      },
      "CheckoutVerifyRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "CheckoutVerifyResponse": {
        "type": "object",
        "additionalProperties": true
      },
      "ConvexProxyRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "ConvexProxyResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}